+++ src/components/MapGL/index.js
@@ -3,12 +3,15 @@
 import { Children, PureComponent, createElement, cloneElement } from 'react';
 import { hash, isImmutable } from 'immutable';
 import type { Node } from 'react';
+import type { EventProps } from './eventProps';
 
 import Layer from '../Layer';
 import MapContext from '../MapContext';
 import mapboxgl from '../../utils/mapbox-gl';
+import events from './events';
+import capitalizeFirstLetter from '../../utils/capitalizeFirstLetter';
 
-type Props = {
+type Props = EventProps & {
   /** container className */
   className?: string,
 
@@ -177,7 +180,10 @@ type Props = {
   onViewportChange?: (viewport: Viewport) => void,
 
   /** The onLoad callback for the map */
-  onLoad?: Function
+  onLoad?: Function,
+
+  /** Map cursor style as CSS value */
+  cursorStyle?: string
 };
 
 type State = {
@@ -192,6 +198,7 @@ class MapGL extends PureComponent<Props, State> {
   static displayName = 'MapGL';
 
   static defaultProps = {
+    children: null,
     className: null,
     style: null,
     mapStyle: 'mapbox://styles/mapbox/light-v8',
@@ -200,6 +207,7 @@ class MapGL extends PureComponent<Props, State> {
     pitch: 0,
     minZoom: 0,
     maxZoom: 22,
+    maxBounds: null,
     hash: false,
     bearingSnap: 7,
     pitchWithRotate: true,
@@ -221,7 +229,8 @@ class MapGL extends PureComponent<Props, State> {
     onLoad: null,
     localIdeographFontFamily: null,
     transformRequest: null,
-    collectResourceTiming: false
+    collectResourceTiming: false,
+    cursorStyle: null
   };
 
   constructor(props: Props) {
@@ -294,6 +303,17 @@ class MapGL extends PureComponent<Props, State> {
       map.on('boxzoomend', this._onViewportChange);
     }
 
+    events.forEach(event => {
+      const propName = `on${capitalizeFirstLetter(event)}`;
+      if (this.props[propName]) {
+        map.on(event, this.props[propName]);
+      }
+    });
+
+    if (this.props.cursorStyle) {
+      map.getCanvas().style.cursor = this.props.cursorStyle;
+    }
+
     this._map = map;
     this._updateMapViewport(this.props);
   }
@@ -301,6 +321,10 @@ class MapGL extends PureComponent<Props, State> {
   componentWillReceiveProps(newProps: Props) {
     this._updateMapViewport(newProps);
     this._updateMapStyle(this.props, newProps);
+
+    if (!newProps.cursorStyle !== this.props.cursorStyle) {
+      this._map.getCanvas().style.cursor = newProps.cursorStyle;
+    }
   }
 
   componentWillUnmount() {